api/requirements.txt
flask
pandas
sentence-transformers
faiss-cpu
numpy
pypdf